requirements.txt
##############################################
# PyTorch (CUDA 12.1) – INSTALL SEPARATELY
##############################################
# ⚠️ Do NOT uncomment these here.
# Install manually using:
# python -m pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
# torch==2.5.1+cu121
# torchvision==0.20.1+cu121
# torchaudio==2.5.1+cu121


##############################################
# Ultralytics YOLO (Core)
##############################################
ultralytics>=8.1.0
numpy>=1.23
opencv-python>=4.8          # use opencv-python (NOT headless) for local + video
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.64.0
typing-extensions>=4.0


##############################################
# Tracking (DeepSort)
##############################################
deep-sort-realtime>=1.3.2


##############################################
# FastAPI / Web (OPTIONAL – safe to keep)
##############################################
fastapi>=0.110
uvicorn[standard]>=0.29
websockets>=12.0
httpx>=0.27
python-multipart>=0.0.9


##############################################
# Cloud / Database Clients (OPTIONAL)
##############################################
psycopg2-binary==2.9.9
boto3==1.34.90
botocore==1.34.90


##############################################
# Dev / Debug Tools
##############################################
flake8>=7.0
